refactor(direct-control): extract App UI snapshot wrapper
Refs: openspec/changes/civ7-support-direct-control-modularization

Move only the App UI snapshot wrapper orchestration, generated App UI snapshot command, and command-result parser into packages/civ7-direct-control/src/runtime/app-ui-snapshot.ts while keeping the public facade export and lifecycle/setup orchestration in packages/civ7-direct-control/src/index.ts.

Restart and setup lifecycle loops continue to reuse the same internal snapshot builder/parser helpers without exposing them through the package facade. Raw App UI snapshot output remains debug/internal service projection; normal CLI projection is omitted/debug-only or left for later semantic state-machine surfaces.

Tuner health, playable status, bounded root inspection, capability catalog, telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, and Effect/oRPC procedure-core work remain pending.

Verification:
- bun run --cwd packages/civ7-direct-control test -- runtime-and-catalog.test.ts restart-lifecycle.test.ts setup-and-lifecycle.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- git diff --check
- ownership and no-cycle scans for App UI snapshot helpers

Local read-only package/source relocation proof only; no runtime or live-game proof claimed.
