+++ packages/civ7-direct-control/src/index.ts
@@ -26,6 +26,11 @@ import {
   parseCiv7TunerFrame,
   type Civ7TunerFrame,
 } from "./session/framing.js";
+import {
+  appUiSnapshotFromCommandResult,
+  buildAppUiSnapshotCommand,
+  getCiv7AppUiSnapshot as getCiv7AppUiSnapshotFromModule,
+} from "./runtime/app-ui-snapshot.js";
 import { inspectCiv7RuntimeApi as inspectCiv7RuntimeApiFromModule } from "./runtime/inspection.js";
 import {
   getCiv7NotificationDismissal as getCiv7NotificationDismissalFromModule,
@@ -2375,11 +2380,9 @@ export async function inspectCiv7RuntimeApi(options: Civ7DirectControlOptions & 
 export async function getCiv7AppUiSnapshot(
   options: Civ7DirectControlOptions = {},
 ): Promise<Civ7AppUiSnapshotResult> {
-  const result = await executeCiv7AppUiCommand({
-    ...options,
-    command: buildAppUiSnapshotCommand(),
+  return await getCiv7AppUiSnapshotFromModule(options, {
+    executeAppUiCommand: executeCiv7AppUiCommand,
   });
-  return appUiSnapshotFromCommandResult(result);
 }
 
 export async function beginCiv7Game(options: Civ7DirectControlOptions = {}): Promise<Civ7CommandResult> {
@@ -4259,87 +4262,6 @@ function portFromEnv(env: NodeJS.ProcessEnv): number | undefined {
 
 function splitEnvList(value: string | undefined): string[] {
   return value?.split(",").map((entry) => entry.trim()).filter(Boolean) ?? [];
-}
-
-function buildAppUiSnapshotCommand(): string {
-  return `(() => {
-    const g = globalThis;
-    const probe = (fn) => {
-      try {
-        return { ok: true, value: fn() };
-      } catch (err) {
-        return { ok: false, error: String(err) };
-      }
-    };
-    const probeValueOr = (fallback, fn) => {
-      const result = probe(fn);
-      return result.ok ? result.value : fallback;
-    };
-    return JSON.stringify({
-      network: {
-        isInSession: probe(() => g.Network.isInSession),
-        numPlayers: probe(() => g.Network.getNumPlayers()),
-        hostPlayerId: probe(() => g.Network.getHostPlayerId()),
-        isConnectedToNetwork: probe(() => g.Network.isConnectedToNetwork()),
-        isAuthenticated: probe(() => g.Network.isAuthenticated()),
-        isLoggedIn: probe(() => g.Network.isLoggedIn()),
-      },
-      autoplay: {
-        isActive: probeValueOr(false, () => typeof g.Autoplay !== "undefined" ? g.Autoplay.isActive : false),
-        turns: probeValueOr(0, () => typeof g.Autoplay !== "undefined" ? g.Autoplay.turns : 0),
-        isPaused: probeValueOr(false, () => typeof g.Autoplay !== "undefined" ? g.Autoplay.isPaused : false),
-        isPausedOrPending: probeValueOr(false, () => typeof g.Autoplay !== "undefined" ? g.Autoplay.isPausedOrPending : false),
-        observeAsPlayer: probeValueOr(-1, () => typeof g.Autoplay !== "undefined" ? g.Autoplay.observeAsPlayer : -1),
-        returnAsPlayer: probeValueOr(-1, () => typeof g.Autoplay !== "undefined" ? g.Autoplay.returnAsPlayer : -1),
-      },
-      game: {
-        turn: probeValueOr(-1, () => g.Game.turn),
-        age: probeValueOr(-1, () => g.Game.age),
-        maxTurns: probeValueOr(0, () => g.Game.maxTurns),
-        turnDate: probe(() => g.Game.getTurnDate()),
-        hash: probe(() => g.Game.getHash()),
-      },
-      ui: {
-        inGame: probe(() => g.UI.isInGame()),
-        inShell: probe(() => g.UI.isInShell()),
-        inLoading: probe(() => g.UI.isInLoading()),
-        loadingState: probe(() => g.UI.getGameLoadingState()),
-        loadingStateName: (() => {
-          try {
-            const state = g.UI.getGameLoadingState();
-            return Object.entries(g.UIGameLoadingState).find(([, value]) => value === state)?.[0] ?? null;
-          } catch {
-            return null;
-          }
-        })(),
-        canBeginGame: probe(() => {
-          const state = g.UI.getGameLoadingState();
-          return state === g.UIGameLoadingState.WaitingForUIReady || state === g.UIGameLoadingState.WaitingToStart;
-        }),
-        canNotifyUIReady: typeof g.UI?.notifyUIReady,
-        skipStartButton: probe(() => g.Configuration.getGame().skipStartButton),
-        automationActive: probe(() => typeof g.Automation !== "undefined" ? g.Automation.isActive : false),
-      },
-      gameContext: {
-        localPlayerID: probeValueOr(-1, () => g.GameContext.localPlayerID),
-        localObserverID: probeValueOr(-1, () => g.GameContext.localObserverID),
-        hasRequestedPause: probe(() => g.GameContext.hasRequestedPause()),
-      },
-      players: {
-        maxPlayers: probeValueOr(0, () => g.Players.maxPlayers),
-        aliveIds: probe(() => g.Players.getAliveIds()),
-        aliveHumanIds: probe(() => g.Players.getAliveHumanIds()),
-        numAliveHumans: probe(() => g.Players.getNumAliveHumans()),
-      },
-      map: {
-        width: probe(() => g.GameplayMap.getGridWidth()),
-        height: probe(() => g.GameplayMap.getGridHeight()),
-        plotCount: probe(() => g.GameplayMap.getPlotCount()),
-        mapSize: probe(() => g.GameplayMap.getMapSize()),
-        randomSeed: probe(() => g.GameplayMap.getRandomSeed()),
-      },
-    });
-  })()`;
 }
 
 function buildTunerHealthCommand(): string {
@@ -6556,23 +6478,6 @@ function tunerStatesFromParts(parts: ReadonlyArray<string>): Civ7TunerState[] {
     states.push({ id: parts[i] ?? "", name: parts[i + 1] ?? "" });
   }
   return states;
-}
-
-function appUiSnapshotFromCommandResult(result: Civ7CommandResult): Civ7AppUiSnapshotResult {
-  try {
-    return {
-      host: result.host,
-      port: result.port,
-      state: result.state,
-      snapshot: JSON.parse(result.output[0] ?? "{}") as Civ7AppUiSnapshot,
-    };
-  } catch (err) {
-    throw new Civ7DirectControlError(
-      "command-failed",
-      `Civ7 App UI snapshot returned invalid JSON: ${result.output.join("\n") || "<empty>"}`,
-      { cause: err, details: result },
-    );
-  }
 }
 
 function tunerHealthFromCommandResult(result: Civ7CommandResult): Civ7TunerHealthResult {
